--- config/config.json
@@ -1,81 +1,0 @@
-{
-  "cohortComparisonResultsEnabled": true,
-  "useExecutionEngine": true,
-  "plpResultsEnabled": false,
-  "webAPIRoot": "http://localhost:8089/WebAPI/",
-  "api": {
-    "name": "Local",
-    "url": "http://localhost:8089/WebAPI/"
-  },
-  "supportUrl": "https://github.com/ohdsi/atlas/issues",
-  "supportMail": "support@odysseusinc.com",
-  "userAuthenticationEnabled":true,
-  "authProviders": [
-    {
-      "name": "Windows",
-      "url": "user/login/windows",
-      "ajax": true,
-      "icon": "fa fa-windows"
-    },
-    {
-      "name": "Kerberos",
-      "url": "user/login/kerberos",
-      "ajax": true,
-      "icon": "fa fa-windows"
-    },
-    {
-      "name": "OpenID",
-      "url": "user/login/openid",
-      "ajax": false,
-      "icon": "fa fa-openid"
-    },
-    {
-      "name": "Google",
-      "url": "user/oauth/google",
-      "ajax": false,
-      "icon": "fa fa-google"
-    },
-    {
-      "name": "Facebook",
-      "url": "user/oauth/facebook",
-      "ajax": false,
-      "icon": "fa fa-facebook"
-    },
-    {
-      "name": "DB",
-      "url": "user/login/db",
-      "ajax": true,
-      "icon": "fa fa-database",
-      "isUseCredentialsForm":true
-    },
-    {
-      "name": "LDAP",
-      "url": "user/login/ldap",
-      "ajax": true,
-      "icon": "fa fa-cubes",
-      "isUseCredentialsForm": true
-    },
-    {
-      "name": "Active Directory LDAP",
-      "url": "user/login/ad",
-      "ajax": true,
-      "icon": "fa fa-cubes",
-      "isUseCredentialsForm": true
-    }
-  ],
-  "xssOptions": {
-    "whiteList": {
-      "a": ["href", "class"],
-      "span": ["class", "data-bind"],
-      "i": ["class", "id", "aria-hidden"],
-      "div": ["class", "style", "id"],
-      "option": ["value"],
-      "input": ["type", "class"],
-      "ui": ["class"],
-      "path": ["d", "class"],
-      "button": ["type", "class"]
-    },
-    "stripIgnoreTag": true,
-    "stripIgnoreTagBody": ["script"]
-  }
-}
